Removing usgs_functions.cpp. Remaining routines moved to ui_usgs_tab.cpp.
